Make the view-count ablation a command instead of a scratch script
It now produces a headline result, so it needs to be reproducible. `cli views`
scores whichever per-view caches exist, prints the table, and writes JSON;
missing caches are reported with the command that builds them rather than
raising, since a partial sweep is still worth seeing.

The module carries the argument the table makes: agreement falls politely from
0.608 to 0.424 as views are removed while the reconstructions stop being
reconstructions, so the plausibility check belongs beside a geometric quality
metric rather than instead of one. Agreement measures whether the views concur;
it cannot notice that they concur on something the size of a wardrobe.

Runbook gains the step. Six tests, including that every view count divides 12 --
a subset that does not cannot be evenly spaced around the rig.

+++ ggssvt/geometry/pot.py
@@ -122,7 +122,7 @@ def estimate_pot_height(
     if not profile.any():
         return PotEstimate(fallback_m, False, 0, 0, 0.0, "empty volume")
 
-    search_max = min(n - 1 - above_slices, int(round(search_max_m / voxel_size_m)))
+    search_max = min(n - 1 - above_slices, round(search_max_m / voxel_size_m))
     if search_max <= below_slices:
         return PotEstimate(fallback_m, False, 0, 0, 0.0, "search window too small")
 
@@ -158,7 +158,7 @@ def estimate_pot_height(
         return PotEstimate(
             height_m=float(rim * voxel_size_m),
             confident=True,
-            body_voxels=int(round(body)),
+            body_voxels=round(body),
             rim_voxels=int(profile[rim]) if rim < n else 0,
             drop_ratio=ratio,
             reason=(
@@ -170,7 +170,7 @@ def estimate_pot_height(
     return PotEstimate(
         height_m=fallback_m,
         confident=False,
-        body_voxels=int(round(body)),
+        body_voxels=round(body),
         rim_voxels=0,
         drop_ratio=1.0,
         reason="no step found; the profile tapers smoothly, so pot and plant "
